refactor: split into models, asserts and tests
